move jinja2 filters & functions to dedicated jinja2_helpers.py
